requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
allure-pytest==2.5.3
allure-python-commons==2.5.3  # via allure-pytest
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via allure-python-commons, pytest
backports.functools-lru-cache==1.5  # via selene
certifi==2018.10.15       # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
colorama==0.4.0           # via crayons
configparser==3.5.0       # via webdriver-manager
crayons==0.1.2            # via webdriver-manager
future==0.17.0            # via selene
idna==2.7                 # via requests
logbook==1.4.1
more-itertools==4.3.0     # via pytest
pip-tools==3.1.0
pluggy==0.8.0             # via allure-python-commons, pytest
py==1.7.0                 # via pytest
pytest-assume==1.2
pytest-dependency==0.2
pytest-ordering==0.6
pytest==3.7.4
requests==2.20.0          # via webdriver-manager
selene==1.0.0a12
selenium==2.53.1          # via selene
six==1.11.0               # via allure-pytest, allure-python-commons, more-itertools, pip-tools, pytest, selene
urllib3==1.24             # via requests
webdriver-manager==1.4.5  # via selene
